aws-cdk/cdk-embed-lambda/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
aws-cdk-lib==2.108.1
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
aws-cdk.aws-lambda-python-alpha==2.108.1a0
boto3==1.29.0
botocore==1.32.0
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.3.2
constructs==10.3.0
dataclasses-json==0.6.2
exceptiongroup==1.1.3
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
importlib-resources==6.1.1
jmespath==1.0.1
jsii==1.91.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.335
langsmith==0.0.64
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
publication==0.0.3
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
